Refactor coverage and move to own module
