+++ infra/cifuzz/cifuzz.py
@@ -18,7 +18,6 @@ This module helps CI tools do the following:
 Eventually it will be used to help CI tools determine which fuzzers to run.
 """
 
-import json
 import logging
 import os
 import shutil
@@ -28,6 +27,7 @@ import urllib.error
 import urllib.request
 
 import fuzz_target
+import coverage
 
 # pylint: disable=wrong-import-position,import-error
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
@@ -69,9 +69,6 @@ STACKTRACE_END_MARKERS = [
 # Default fuzz configuration.
 DEFAULT_ENGINE = 'libfuzzer'
 DEFAULT_ARCHITECTURE = 'x86_64'
-
-# The path to get project's latest report json files.
-LATEST_REPORT_INFO_PATH = 'oss-fuzz-coverage/latest_report_info/'
 
 # TODO(metzman): Turn default logging to WARNING when CIFuzz is stable.
 logging.basicConfig(
@@ -426,7 +423,7 @@ def run_fuzzers(  # pylint: disable=too-many-arguments,too-many-locals
     logging.error('Invalid workspace: %s.', workspace)
     return False, False
 
-  logging.info("Using %s sanitizer.", sanitizer)
+  logging.info('Using %s sanitizer.', sanitizer)
 
   out_dir = os.path.join(workspace, 'out')
   artifacts_dir = os.path.join(out_dir, 'artifacts')
@@ -531,95 +528,6 @@ def check_fuzzer_build(out_dir,
   return True
 
 
-def get_latest_cov_report_info(project_name):
-  """Gets latest coverage report info for a specific OSS-Fuzz project from GCS.
-
-  Args:
-    project_name: The name of the relevant OSS-Fuzz project.
-
-  Returns:
-    The projects coverage report info in json dict or None on failure.
-  """
-  latest_report_info_url = fuzz_target.url_join(fuzz_target.GCS_BASE_URL,
-                                                LATEST_REPORT_INFO_PATH,
-                                                project_name + '.json')
-  latest_cov_info_json = get_json_from_url(latest_report_info_url)
-  if not latest_cov_info_json:
-    logging.error('Could not get the coverage report json from url: %s.',
-                  latest_report_info_url)
-    return None
-  return latest_cov_info_json
-
-
-def get_target_coverage_report(latest_cov_info, target_name):
-  """Get the coverage report for a specific fuzz target.
-
-  Args:
-    latest_cov_info: A dict containing a project's latest cov report info.
-    target_name: The name of the fuzz target whose coverage is requested.
-
-  Returns:
-    The targets coverage json dict or None on failure.
-  """
-  if 'fuzzer_stats_dir' not in latest_cov_info:
-    logging.error('The latest coverage report information did not contain'
-                  '\'fuzzer_stats_dir\' key.')
-    return None
-  fuzzer_report_url_segment = latest_cov_info['fuzzer_stats_dir']
-
-  # Converting gs:// to http://
-  fuzzer_report_url_segment = fuzzer_report_url_segment.replace('gs://', '')
-  target_url = fuzz_target.url_join(fuzz_target.GCS_BASE_URL,
-                                    fuzzer_report_url_segment,
-                                    target_name + '.json')
-  return get_json_from_url(target_url)
-
-
-def get_files_covered_by_target(latest_cov_info, target_name,
-                                oss_fuzz_repo_path):
-  """Gets a list of source files covered by the specific fuzz target.
-
-  Args:
-    latest_cov_info: A dict containing a project's latest cov report info.
-    target_name: The name of the fuzz target whose coverage is requested.
-    oss_fuzz_repo_path: The location of the repo in the docker image.
-
-  Returns:
-    A list of files that the fuzzer covers or None.
-  """
-  if not oss_fuzz_repo_path:
-    logging.error('Project source location in docker is not found.'
-                  'Can\'t get coverage information from OSS-Fuzz.')
-    return None
-  target_cov = get_target_coverage_report(latest_cov_info, target_name)
-  if not target_cov:
-    return None
-  coverage_per_file = target_cov['data'][0]['files']
-  if not coverage_per_file:
-    logging.info('No files found in coverage report.')
-    return None
-
-  # Make sure cases like /src/curl and /src/curl/ are both handled.
-  norm_oss_fuzz_repo_path = os.path.normpath(oss_fuzz_repo_path)
-  if not norm_oss_fuzz_repo_path.endswith('/'):
-    norm_oss_fuzz_repo_path += '/'
-
-  affected_file_list = []
-  for file in coverage_per_file:
-    norm_file_path = os.path.normpath(file['filename'])
-    if not norm_file_path.startswith(norm_oss_fuzz_repo_path):
-      continue
-    if not file['summary']['regions']['count']:
-      # Don't consider a file affected if code in it is never executed.
-      continue
-
-    relative_path = file['filename'].replace(norm_oss_fuzz_repo_path, '')
-    affected_file_list.append(relative_path)
-  if not affected_file_list:
-    return None
-  return affected_file_list
-
-
 def remove_unaffected_fuzzers(project_name, out_dir, files_changed,
                               oss_fuzz_repo_path):
   """Removes all non affected fuzzers in the out directory.
@@ -638,16 +546,13 @@ def remove_unaffected_fuzzers(project_name, out_dir, files_changed,
     logging.error('No fuzzers found in out dir.')
     return
 
-  latest_cov_report_info = get_latest_cov_report_info(project_name)
-  if not latest_cov_report_info:
-    logging.error('Could not download latest coverage report.')
-    return
+  coverage_getter = coverage.OSSFuzzCoveragGetter(
+      project_name, oss_fuzz_repo_path)
   affected_fuzzers = []
   logging.info('Files changed in PR:\n%s', '\n'.join(files_changed))
   for fuzzer in fuzzer_paths:
     fuzzer_name = os.path.basename(fuzzer)
-    covered_files = get_files_covered_by_target(latest_cov_report_info,
-                                                fuzzer_name, oss_fuzz_repo_path)
+    covered_files = coverage_getter.get_files_covered_by_target(fuzzer_name)
     if not covered_files:
       # Assume a fuzzer is affected if we can't get its coverage from OSS-Fuzz.
       affected_fuzzers.append(fuzzer_name)
@@ -673,29 +578,6 @@ def remove_unaffected_fuzzers(project_name, out_dir, files_changed,
         os.remove(os.path.join(out_dir, fuzzer))
       except OSError as error:
         logging.error('%s occurred while removing file %s', error, fuzzer)
-
-
-def get_json_from_url(url):
-  """Gets a json object from a specified HTTP URL.
-
-  Args:
-    url: The url of the json to be downloaded.
-
-  Returns:
-    A dictionary deserialized from JSON or None on failure.
-  """
-  try:
-    response = urllib.request.urlopen(url)
-  except urllib.error.HTTPError:
-    logging.error('HTTP error with url %s.', url)
-    return None
-  try:
-    # read().decode() fixes compatibility issue with urllib response object.
-    result_json = json.loads(response.read().decode())
-  except (ValueError, TypeError, json.JSONDecodeError) as err:
-    logging.error('Loading json from url %s failed with: %s.', url, str(err))
-    return None
-  return result_json
 
 
 def parse_fuzzer_output(fuzzer_output, out_dir):
